feat: continue impl
